Move read_participants_file function to utils.py
